scx_layered: Fix core allocation for StickyDynamic
Currently, we replace the layer core order and let the normal core
allocation logic (written with the assumption that it operates on a
fixed core order) to allocate and free CPUs for us. However, this does
not work well and can lead to a broken state when layers are routinely
emptied and StickyDynamic is enabled.

The idea of pushing and popping cores in the order defined by the core
order does not work well when it's changing contiguous runs repeatedly
in such situations, and can lead to cores being "locked" to an emptied
layer which leads to failure in core allocation.

Instead, we need to compute the precise set of cores to be added and
freed everytime we recompute the layer core order for StickyDynamic, and
modify the layer cpumasks directly instead of going through the normal
logic.

This also improves performance in the general case without options to
empty layers, where previous behavior would lead to fragmentation of a
layer across LLCs instead of core failures, leading to lower performance
than simply packing it to a full LLC where possible.

Minimize the fallout of this change by restricting this to
StickyDynamic, we don't want to change the behavior for other growth
algorithms, since any change in behavior causes performance differences.

Overall, this change allows HintEquals modifiers (SystemCpuUtilBelow and
DsqUtilBelow) to function correctly when emptying layers, and also
provides a performance improvement at saturation for production
workloads.

Fixes: fc4e68dbbaa3 ("scx_layered: Introduce StickyDynamic growth algorithm")
Signed-off-by: Kumar Kartikeya Dwivedi <memxor@gmail.com>